+++ zic/zic_wavetable_base.h
@@ -4,11 +4,11 @@
 class Zic_Wavetable_Base
 {
 public:
-    float *table;
+    int16_t *table;
     uint16_t size;
     uint16_t sampleCount;
 
-    Zic_Wavetable_Base(float *_table, uint16_t _sampleCount, uint16_t _size)
+    Zic_Wavetable_Base(int16_t *_table, uint16_t _sampleCount, uint16_t _size)
     {
         table = _table;
         size = _size;
